handled basic string get set del and connection quit operations

+++ src/query.c
@@ -61,45 +61,82 @@ int parse_statefull_request(dstring* requestSequence, query* query_p)
 
 int process_query(dstring* responseSequence, query* query_p)
 {
+	make_dstring_empty(responseSequence);
+
 	int exit_called = 0;
+
+	int delete_query_key = 0;
+	int delete_query_value = 0;
+
+
 	switch(query_p->command)
 	{
 		case GET :
 		{
-			printf("GET\nKEY : ");
-			display_dstring(query_p->key);printf("\n\n");
+			dstring* value = (dstring*)(find_value_from_hash(hashTable, query_p->key));
+			delete_query_key = 1;
+			delete_query_value = 0;
+			if(value != NULL)
+			{
+				responseSequence = append_to_dstring(responseSequence, "\"");
+				responseSequence = concatenate_dstring(responseSequence, value);
+				responseSequence = append_to_dstring(responseSequence, "\"");
+			}
+			else
+			{
+				responseSequence = append_to_dstring(responseSequence, "NULL");
+			}
+			responseSequence = append_to_dstring(responseSequence, "\n");
 			break;
 		}
 		case SET :
 		{
-			printf("SET\nKEY : ");
-			display_dstring(query_p->key);printf("\nVAL : ");
-			display_dstring(query_p->key);printf("\n\n");
+			dstring* value = (dstring*)(find_value_from_hash(hashTable, query_p->key));
+			if(value == NULL)
+			{
+				insert_entry_in_hash(hashTable, query_p->key, query_p->value);
+				delete_query_key = 0;
+				delete_query_value = 0;
+				responseSequence = append_to_dstring(responseSequence, "ENTRY INSERTED\n");
+			}
+			else
+			{
+				make_dstring_empty(value);concatenate_dstring(value, query_p->value);
+				delete_query_key = 1;
+				delete_query_value = 1;
+				responseSequence = append_to_dstring(responseSequence, "ENTRY UPDATED\n");
+			}
 			break;
 		}
 		case DEL :
 		{
-			printf("DEL\nKEY : ");
-			display_dstring(query_p->key);printf("\n\n");
+			dstring* return_key;
+			dstring* return_value;
+			delete_entry_from_hash(hashTable, query_p->key, (const void**)(&return_key), (const void**)(&return_value));
+			delete_query_key = 1;
+			delete_query_value = 0;
+			responseSequence = append_to_dstring(responseSequence, "ENTRY DELETED\n");
+			delete_dstring(return_key);
+			delete_dstring(return_value);
 			break;
 		}
 		case EXIT :
 		default :
 		{
-			printf("EXIT\n");
 			exit_called = 1;
 			break;
 		}
 	}
 
-	// delete query dstrings
-	if(query_p->key != NULL)
+	if(delete_query_key && query_p->key != NULL)
 	{
 		delete_dstring(query_p->key);
+		query_p->key = NULL;
 	}
-	if(query_p->value != NULL)
+	if(delete_query_value && query_p->value != NULL)
 	{
 		delete_dstring(query_p->value);
+		query_p->value = NULL;
 	}
 
 	return exit_called;
